+++ packages/cleanly_architected_core/lib/src/data_source/remote_data_source.dart
@@ -17,12 +17,13 @@ abstract class RemoteQueryDataSource<T extends EquatableEntity,
   Future<List<T>> read({int pageSize, int pageNumber, U params});
 }
 
-/// The data source which responsible to mutate data from the remote client.
-/// Mutation is a term to change data. So it includes create and update.
+/// The data source which responsible to handle form related request to the
+/// server (create and update).
+///
 /// If you don't need a particular function, just throw a
 /// [CleanException] when calling that function.
 abstract class RemoteMutationDataSource<T extends EquatableEntity,
-    U extends MutationParams<T>> {
+    U extends FormParams<T>> {
   /// Api client which interfaced with [CleanApiClient]. Feel free to use your
   /// own abstract and put it in this class's implementation
   final CleanApiClient client;
@@ -34,17 +35,7 @@ abstract class RemoteMutationDataSource<T extends EquatableEntity,
 
   /// Update data to [params] and return T as the result.
   Future<T> update({@required U params});
-}
 
-/// The data source which responsible to delete data from the remote client.
-abstract class RemoteDeletionDataSource<T extends EquatableEntity,
-    U extends DeletionParams<T>> {
-  // Api client which interfaced with [CleanApiClient]. Feel free to use your
-  /// own abstract and put it in this class's implementation
-  final CleanApiClient client;
-
-  const RemoteDeletionDataSource({this.client});
-
-  /// Delete a data with satisfies specified [params].
-  Future<void> delete({@required U params});
+  /// Update data in server with [id]
+  Future<void> delete({String id});
 }