+++ packages/cleanly_architected_core/lib/src/interactor/read_interactors.dart
@@ -1,17 +1,17 @@
 import 'package:cleanly_architected_core/src/entity/clean_error.dart';
 import 'package:cleanly_architected_core/src/entity/equatable_entity.dart';
-import 'package:cleanly_architected_core/src/repository/query_repository.dart';
+import 'package:cleanly_architected_core/src/repository/data_repository.dart';
 import 'package:cleanly_architected_core/src/data_source/params.dart';
 import 'package:dartz/dartz.dart';
 import 'package:meta/meta.dart';
 
-/// Base interactor to call ReadNext.
+/// Base readNext interactor / use case.
 ///
 /// This doesn't perform any validation.
 /// If you need to validate, feel free to extends this. Otherwise, just register
 /// it with T.
 class ReadNext<T extends EquatableEntity, U extends QueryParams<T>> {
-  final QueryRepository<T, U> repo;
+  final DataRepository<T, U> repo;
 
   const ReadNext({@required this.repo});
 
@@ -24,13 +24,13 @@ class ReadNext<T extends EquatableEntity, U extends QueryParams<T>> {
           pageNumber: pageNumber, pageSize: pageSize, params: params);
 }
 
-/// Base interactor to call RefreshAll.
+/// Base refreshAll interactor / use case.
 ///
 /// This doesn't perform any validation.
 /// If you need to validate, feel free to extends this. Otherwise, just register
 /// it with T.
 class RefreshAll<T extends EquatableEntity, U extends QueryParams<T>> {
-  final QueryRepository<T, U> repo;
+  final DataRepository<T, U> repo;
 
   const RefreshAll({@required this.repo});