+++ packages/cleanly_architected_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.4] - 12 December 2020
+- Rework architecture.
+- Interactor / Use case or `Create`, `Update` and `Delete` will automatically reflects the change to `DataRepository` as well.
+- [TODO]: Empty form_cache_repository to be made later.
+
 ## [0.0.3] - 11 December 2020
 - Add coverage badge to README.
 - Format some document to pass pub.dev static analysis.